Lab3_AI_Agent/requirements.txt
scapy>=2.5.0
requests>=2.31.0
google-generativeai>=0.7.0
